fix: 修复目标速度因数据类型转换而导致的数值错误

+++ code/control/guide/guide.c
@@ -5,21 +5,21 @@
 #include "detection.h"
 #include "pid.h"
 
-static int16 guide_target_vel = 0;
-static int16 guide_target_turn = 0;
+static int32 guide_target_vel = 0;
+static int32 guide_target_turn = 0;
 
 static int16 last_target_error = 0;
 static bool target_lost = false;
 static int16 search_turn_value = 1000; // 目标丢失时的转向强度
 
-void guide_set_target_vel(int16 target_vel)
+void guide_set_target_vel(int32 target_vel)
 {
     restrictValueI(&target_vel, -9999, 9999);
 
     guide_target_vel = -target_vel;
 }
 
-void guide_set_target_turn(int16 target_turn)
+void guide_set_target_turn(int32 target_turn)
 {
     // TODO 限幅
     guide_target_turn = target_turn;
